Revert "making changes towards beve"
This reverts commit 6f94d95ce5e7b01ce9b7723331cc1da577be08fc.
