+++ libs/ipc/inc/public/tfc/ipc/enums.hpp
@@ -5,10 +5,6 @@
 #include <string_view>
 #include <utility>
 
-#include <glaze/core/common.hpp>
-
-#include <tfc/stx/basic_fixed_string.hpp>
-
 // Common enums, ipc/glaze_meta.hpp provides glaze specific conversions
 
 namespace tfc::ipc::details {
@@ -88,44 +84,3 @@ static_assert(enum_name(type_e::_string) == "string");
 static_assert(enum_name(type_e::_json) == "json");
 
 }  // namespace tfc::ipc::details
-
-
-namespace tfc::ipc {
-
-inline stx::basic_fixed_string constexpr signal_tag{ "signal" };
-inline stx::basic_fixed_string constexpr slot_tag{ "slot" };
-
-}  // namespace tfc::ipc
-
-namespace glz {
-
-template <>
-struct meta<tfc::ipc::details::direction_e> {
-  using enum tfc::ipc::details::direction_e;
-  // clang-format off
-  static auto constexpr value{ glz::enumerate("unknown", unknown, "Unspecified direction",
-                                              tfc::ipc::signal_tag.data_, signal, "Owner of information being sent/published",
-                                              tfc::ipc::slot_tag.data_, slot, "Receiver of information, or subscriber")
-  };
-  // clang-format on
-  static std::string_view constexpr name{ "direction_e" };
-};
-
-template <>
-struct meta<tfc::ipc::details::type_e> {
-  using enum tfc::ipc::details::type_e;
-  // clang-format off
-  static auto constexpr value{ glz::enumerate(
-    tfc::ipc::details::type_e_iterable[std::to_underlying(unknown)], unknown, "Unspecified type",
-    tfc::ipc::details::type_e_iterable[std::to_underlying(_bool)], _bool, "Boolean",
-    tfc::ipc::details::type_e_iterable[std::to_underlying(_int64_t)], _int64_t, "Signed 64bit integer",
-    tfc::ipc::details::type_e_iterable[std::to_underlying(_uint64_t)], _uint64_t, "Unsigned 64bit integer",
-    tfc::ipc::details::type_e_iterable[std::to_underlying(_double_t)], _double_t, "Double",
-    tfc::ipc::details::type_e_iterable[std::to_underlying(_string)], _string, "String",
-    tfc::ipc::details::type_e_iterable[std::to_underlying(_json)], _json, "Json",
-    tfc::ipc::details::type_e_iterable[std::to_underlying(_mass)], _mass, "Mass"
-  ) };
-  // clang-format on
-};
-
-}  // namespace glz
